+++ R/monitorPlot_timeseries.R
@@ -76,6 +76,11 @@ monitorPlot_timeseries <- function(ws_monitor,
     hourInterval=6
   }
   
+  # Sanity check
+  if ( monitor_isEmpty(ws_monitor) ) {
+    stop("ws_monitor object contains zero monitors")
+  }
+  
   # ----- Data Preparation ------------
   
   # Set monitorID if not passed in as argument