+++ R/monitor_isolate.R
@@ -27,9 +27,16 @@
 #' monitorList <- monitor_isolate(Spokane)
 #' names(monitorList)
 
-monitor_isolate <- function(ws_monitor, xlim=NULL, ylim=NULL, tlim=NULL,
-                            monitorIDs=NULL, stateCodes=NULL,
+monitor_isolate <- function(ws_monitor,
+                            xlim=NULL,
+                            ylim=NULL,
+                            tlim=NULL,
+                            monitorIDs=NULL,
+                            stateCodes=NULL,
                             timezone="UTC") {
+  
+  # Sanity check
+  if ( monitor_isEmpty(ws_monitor) ) stop("ws_monitor object contains zero monitors")
   
   # Isolate individual monitors
   monList <- list()